+++ extension/android/executorch_android/src/main/java/org/pytorch/executorch/extension/audio/package-info.java
@@ -1,2 +1,2 @@
-/** Extension for ASR related use cases for ExecuTorch Android Java/JNI package. */
+/** Extension for audio and ASR related use cases for ExecuTorch Android Java/JNI package. */
 package org.pytorch.executorch.extension.audio;
